Split main.dart file.
